bot/requirements.txt
asyncio==3.4.3
motor~=3.3.1
pydantic~=2.4.2
pymongo==4.6.1
cachetools~=5.3.0
python-dotenv==1.0.0
aiogram==3.3.0